Analog

+++ train-controller/src/main/java/hu/bme/mit/train/controller/TrainControllerImpl.java
@@ -4,9 +4,9 @@ import hu.bme.mit.train.interfaces.TrainController;
 
 public class TrainControllerImpl implements TrainController {
 
-	private int step = 0;
-	private int referenceSpeed = 0;
-	private int speedLimit = 0;
+	private double step = 0;
+	private double referenceSpeed = 0;
+	private double speedLimit = 0;
 
 	@Override
 	public void followSpeed() {
@@ -20,12 +20,12 @@ public class TrainControllerImpl implements TrainController {
 	}
 
 	@Override
-	public int getReferenceSpeed() {
+	public double getReferenceSpeed() {
 		return referenceSpeed;
 	}
 
 	@Override
-	public void setSpeedLimit(int speedLimit) {
+	public void setSpeedLimit(double speedLimit) {
 		this.speedLimit = speedLimit;
 		enforceSpeedLimit();
 		
@@ -38,7 +38,7 @@ public class TrainControllerImpl implements TrainController {
 	}
 
 	@Override
-	public void setJoystickPosition(int joystickPosition) {
+	public void setJoystickPosition(double joystickPosition) {
 		this.step = joystickPosition;		
 	}
